refactor(test): migrate module specs from SQLite to testcontainers
