+++ apps/api/src/courses/courses.module.spec.ts
@@ -1,40 +1,16 @@
 import { Test, TestingModule } from "@nestjs/testing";
 import { TypeOrmModule } from "@nestjs/typeorm";
-import { Association } from "@/associations/entities/association.entity";
-import { Event } from "@/events/entities/event.entity";
-import { Faculty } from "@/faculties/entities/faculty.entity";
-import { Request } from "@/requests/entities/request.entity";
-import { Schedule } from "@/services/entity/schedule.entity";
-import { Service } from "@/services/entity/service.entity";
-import { User } from "@/users/entities/user.entity";
+import { getTestDatabaseOptions } from "@/test/test-database";
 import { CoursesController } from "./courses.controller";
 import { CoursesModule } from "./courses.module";
 import { CoursesService } from "./courses.service";
-import { Course } from "./entities/course.entity";
 
 describe("CoursesModule", () => {
   let module: TestingModule;
 
   beforeAll(async () => {
     module = await Test.createTestingModule({
-      imports: [
-        TypeOrmModule.forRoot({
-          type: "sqlite",
-          database: ":memory:",
-          entities: [
-            Association,
-            Course,
-            Faculty,
-            Event,
-            Service,
-            Schedule,
-            User,
-            Request,
-          ],
-          synchronize: true,
-        }),
-        CoursesModule,
-      ],
+      imports: [TypeOrmModule.forRoot(getTestDatabaseOptions()), CoursesModule],
     }).compile();
   });
