+++ apps/api/src/requests/requests.module.spec.ts
@@ -1,15 +1,8 @@
 import { Test, TestingModule } from "@nestjs/testing";
 import { TypeOrmModule } from "@nestjs/typeorm";
-import { Association } from "@/associations/entities/association.entity";
-import { Course } from "@/courses/entities/course.entity";
-import { Event } from "@/events/entities/event.entity";
 import { EventsService } from "@/events/events.service";
-import { Faculty } from "@/faculties/entities/faculty.entity";
-import { Schedule } from "@/services/entity/schedule.entity";
-import { Service } from "@/services/entity/service.entity";
 import { ServicesService } from "@/services/services.service";
-import { User } from "@/users/entities/user.entity";
-import { Request } from "./entities/request.entity";
+import { getTestDatabaseOptions } from "@/test/test-database";
 import { RequestsController } from "./requests.controller";
 import { RequestsModule } from "./requests.module";
 import { RequestsService } from "./requests.service";
@@ -20,21 +13,7 @@ describe("RequestsModule", () => {
   beforeAll(async () => {
     module = await Test.createTestingModule({
       imports: [
-        TypeOrmModule.forRoot({
-          type: "sqlite",
-          database: ":memory:",
-          entities: [
-            Association,
-            Course,
-            Faculty,
-            Event,
-            Service,
-            Schedule,
-            User,
-            Request,
-          ],
-          synchronize: true,
-        }),
+        TypeOrmModule.forRoot(getTestDatabaseOptions()),
         RequestsModule,
       ],
     }).compile();
