+++ apps/api/src/users/users.module.spec.ts
@@ -1,5 +1,6 @@
 import { Test, TestingModule } from "@nestjs/testing";
 import { TypeOrmModule } from "@nestjs/typeorm";
+import { getTestDatabaseOptions } from "@/test/test-database";
 import { UsersController } from "./users.controller";
 import { UsersModule } from "./users.module";
 import { UsersService } from "./users.service";
@@ -9,15 +10,7 @@ describe("UsersModule", () => {
 
   beforeAll(async () => {
     module = await Test.createTestingModule({
-      imports: [
-        TypeOrmModule.forRoot({
-          type: "sqlite",
-          database: ":memory:",
-          entities: [`${__dirname}/../**/*.entity{.ts,.js}`],
-          synchronize: true,
-        }),
-        UsersModule,
-      ],
+      imports: [TypeOrmModule.forRoot(getTestDatabaseOptions()), UsersModule],
     }).compile();
   });
